make test components the primary files. start basic hotkey for child component creation
