--- server/db/seeds/guuh_seed.sql
@@ -1,4 +1,0 @@
-INSERT INTO pseudos ( title )
-VALUES ( 'guuuh' );
-
-# mysql -u root -p pseudos < C:\Users\Administrator\Desktop\Coding\webDev\Pseudo\server\db\seeds\guuh_seed.sql
