+++ server/db/seeds/hello_seed.sql
@@ -1,0 +1,4 @@
+INSERT INTO pseudos ( title )
+VALUES ( 'Create function that returns "Hello World!"' );
+
+# mysql -u root -p pseudos < C:\Users\Administrator\Desktop\Coding\webDev\Pseudo\server\db\seeds\hello_seed.sql
