+++ src/App.jsx
@@ -11,7 +11,7 @@ import  CodeMirror from 'react-codemirror';
 import Nav from './components/Nav';
 import Title from './components/Title';
 // import SwitchContainer from './components/SwitchContainer';
-import SwitchContainer from './components/testComponents/SwitchContainerTest';
+import SwitchContainer from './components/SwitchContainer';
 
 
 /////////////////// IMPORTS - CSS //////////////////////////////
