+++ src/components/SwitchContainer.jsx
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import TextEditingContainer from './TextEditingContainer';
-// import PCInputContainer from './PCInputContainer';
-import PCInputContainer from './PCInputContainerTest';
+import PCInputContainer from './PCInputContainer';
 import axios from 'axios';
 
 const EditorContext = React.createContext({
@@ -35,7 +34,8 @@ class SwitchContainer extends Component {
 			fullCode: "",
 			mode: 'javascript',
 			theme: 'midnight',
-			lastPseudo: ''
+			lastPseudo: '',
+			force: true
 		}
 		this.addLastPseudo = this.addLastPseudo.bind(this);
 
@@ -57,16 +57,14 @@ class SwitchContainer extends Component {
 		this.updateStateValue = this.updateStateValue.bind(this);
 	}
 
-	// componentDidMount(){
-	// 	axios.get('/data')
-	// 	.then((response) => {
-	// 		this.setState({
-	// 			pseudos: response
-	// 		});
-	// 	}, () => {
-	// 		console.log(this.state.pseudos)
-	// 	});
-	// }
+	componentWillMount(){
+		axios.get('/data')
+		.then((response) => {
+			this.setState({
+				pseudos: response.data
+			});
+		});
+	}
 
 	/////////////////////////// NAVIGATION ////////////////////////////////
 
@@ -94,29 +92,61 @@ class SwitchContainer extends Component {
 		// if (key === 'Enter'){
 			console.log(pseudo);
 			this.setState(state => { 
-				let pseudos = state.pseudos.concat(pseudo);
-				return { pseudos };
+				state.pseudos.push({
+					title: pseudo
+				});
+			}, () => {
+				console.log('Switch',this.state.pseudos);
+				// return this.state.pseudos;
+				this.setState({
+					force: !this.state.force
+				});
 			});	
 			
+			// this.state.pseudos.push(pseudo)
+
 			// event.target.value = '';
 		// }
 	}
 	
-	editPseudo(index, text){
+	editPseudo(index, newComment, id){
+		console.log(index, newComment, id);
+		// this.setState(state => {
+		// 	state.pseudos.splice(index, 1, newComment);
+		// 	return state.pseudos;
+		// },
 		this.setState(state => {
-			state.pseudos.splice(index, 1, text);
-			return state.pseudos;
+			state.pseudos[index].title = newComment;
+		}, () => {
+			this.setState({
+				force: !this.state.force
+			});
+
+			axios.put(`/data/${id}`, { 
+				title: newComment
+			})
+			.then(response => {
+				console.log(response);
+			})
+			.catch(err => {
+				console.log(err);
+			})
 		})
 	}
 
-    removePseudo(index){
-    		console.log('remove', index);
-			this.setState(state => {
-				let pseudos = state.pseudos.filter((saved, i) => {
-					return i !== index;
-				})
-				return { pseudos }
+    removePseudo(index, id){
+		console.log('remove', index);
+		this.setState(state => {
+			let pseudos = state.pseudos.filter((saved, i) => {
+				return i !== index;
 			})
+			return { pseudos }
+		}, () => {
+			axios.delete(`/data/${id}`)
+			.then((response) => {
+				console.log(response);
+			});
+		})
     }
 
     ///////////////// CENTRALIZED TEXT EDITOR FUNCTIONS ///////////////////
@@ -146,7 +176,7 @@ class SwitchContainer extends Component {
 	}
 
 	addLastPseudo(p){
-		console.log(p);
+		// console.log(p);
 		this.setState({
 			lastPseudo: p
 		})
@@ -159,10 +189,8 @@ class SwitchContainer extends Component {
 		options.mode = this.state.mode;
 		options.theme = this.state.theme;
 
-
 		let switchView = this.state.switchView;
 		let currentView;
-
 
 		if (switchView === 'PC') {
 
@@ -173,13 +201,13 @@ class SwitchContainer extends Component {
 				// 			  </PseudoContext.Provider>
 			//PCInputContext
 			currentView = <PCInputContainer
-											pseudos={this.state.pseudos}
-											editPseudo={this.editPseudo}
-											addPseudo={this.addPseudo}
-											removePseudo={this.removePseudo} 
-											switchToTextEditorView={this.switchToTextEditorView}
-											lastPseudo={this.addLastPseudo}
-			      	  		/>
+							pseudos={this.state.pseudos}
+							editPseudo={this.editPseudo}
+							addPseudo={this.addPseudo}
+							removePseudo={this.removePseudo} 
+							switchToTextEditorView={this.switchToTextEditorView}
+							lastPseudo={this.addLastPseudo}
+			      	  	/>
 		} else if (switchView === 'TE') {
 
 				// currentView = <EditorContext.Provider
@@ -190,20 +218,19 @@ class SwitchContainer extends Component {
 
 			//CodeEditingContext
 				currentView = <TextEditingContainer
-			    	  	    pseudos={this.state.pseudos}
-			    	  	    options={options} 
-			    	  	    fullCode={this.state.fullCode}
-			    	  	    switchToPCView={this.switchToPCView}
-			    	  	    formatComment={this.formatComment}
-			    	  	    constructValueFromProps={this.constructValueFromProps}
-			    	  	    updateStateValue={this.updateStateValue}
+				    	  	    pseudos={this.state.pseudos}
+				    	  	    options={options} 
+				    	  	    fullCode={this.state.fullCode}
+				    	  	    switchToPCView={this.switchToPCView}
+				    	  	    formatComment={this.formatComment}
+				    	  	    constructValueFromProps={this.constructValueFromProps}
+				    	  	    updateStateValue={this.updateStateValue}
 			    	  		/>
 		}
 
 
 		return (
 			<React.Fragment>
-
 				{currentView}
 			</React.Fragment>
 		)
